[Bugfix] Restructured the generator/size/value_type metafunctions gang that were handling expression building.
This restructuration allow for expressions to compute a proper storage type for their size, thus enabling local
optimization on elementwise expressions to not copy size in every nodes on one hand, and, on the other hand,
allow for special storage handling for expressions requiring a computed size to be stored by value.

This also fixes failing test of core.container.function.subscript that was incorrectly using a reference
to temporary in its implementation.