Proposal for faster to compile and more intuitive use of operator() on container.
This patch changes the way expression build function nodes and dispatch it to a
proper implementation for container that returns actual reference to values.
By bypassing the generic operator() indexing resolving system for this very specific
case, this patch fixes the 'cast required' when using operator() on container
and improve compiles times between 10 and 50%.

This patch also contains a fixed tril and subscript unit tests to demonstrate the
change of compile time.