transform the input into a vertical vector if and only if the input is a vector.
Some matlab functions (cov for instance) have a different behaviour
according the fact the input(s) are vector or proper matrices.
This behaviour is generally the same as if the vector was a nx1 matrix.
The functor ifvectvert transform the vectors in vertical ones and don't modify
proper matrices.

This functor is an helper for these functors.