Fix #192 - Extremum indexing returns scalar values
Expressions like `a(end_-n)` or `a(begin_+n)` on any number of dimensions now
yields a scalar instead of an expression.

This is done through a `fix_index` function that pre-evaluate a subset of indexer
before calling the `function_` node constructor, thus redirecting such indexing
to the full scalar indexer instead of the expression based one.

`fix_index` is open for extension if any future indexer need a similar behavior.