ind2sub small correction for fringe cases
ind2sub should not be called on empty expressions. But there is cases where for
example we have to concatenate 2 potentially empty array expressions.
This works well with proper arrays but a small provision must be done if they are
expressions returning empty arrays in order ind2sub doesnot zero divide to obtain a
value that however will not be used.