change of the init for maximum
maximum as any reduction needs an init value that was provided by Valmin,
but in the complex context the order used by max is not the same : it is the
lexicographical order on the pairs (modulus,argument) so the init value must have a zero argument
Maxinit is introduced to do so. (As Valmin for reals but zero for complex)

Note that it is unecessary to make corresponding changes for minimum as Valmax is a good init in every cases.