Introduce the pure_constant_ hierarchy
As a gateway to handle generative functions and constants with the
least replication of code, we introduce the pure_constant_ hierarchy
that flag a constant as beign stateless, thus being able to be
default-constructed from its tag descriptor.

Common handling of constants are now set to look for pure_constant_
hierarchy explicitly.