Most problems related with old (3.2.1) LAPACK version seem solved.
They depended of two bugs in LAPACK 3.2.1 on size queries for svd and schur factoriations.
Doing the alloc "manually" with sufficient (perhaps too large size) solve the problem,
but a version detection it is still needed to insure the use of best allocation values of LAPACK
workspaces when its possible.