maximum in complex case was broken
In complex case maximum use a different ordering on the complex values
than the trivial one used by comparison operators (i.e. comparing real parts)
In fact it compare modulus and then lexically argument
This implies that the max reduction must use a starting value of 0 (not -inf)
because the comparison will begin by some abs(a0) > abs(init)
which will be false with abs(-inf)...

Maxinit is -inf for reals, but the overloding for complex is not included
because boos::simd has no complex. It must be however included somewhere:
for now it's there but we could find a better place than in a file where it is
not explicitely used.