Clean-up and refurbishing of LU decomposition
Implementation of lu now uses an architecture-dependent version
of trf, which is a direct mapping to the underlying LAPACK kernel.
This allows LU to use MAGMA and other specific version of BLAS/LAPACK.

In addition, proper allocation avoiding scheme is now in place
whenever possible for calls of tie(l,u,...) = lu(a). Previous code
was overcopying in some situations (like non-square input).

This change deprecates the factorization namespace as the internals
of the factorization and the memory handling is done by trf or by
specific version of LU itself. Functions depending on other members
of the defunct lu_result type will be rewrote later using trf.