modifying tests for zero crosing rate
In fact zero_crossing rate is devoted to real noisy signal and can be very touchy for inexact values around
zero. But in fact this is not very relevant. The problematc test is related to the fact that cospi(half_int)
can be not exactly computed as 0 depending of the system.