Introduce refine_rec
Fast reciprocate computation is based on callign a raw estimator
followed by some number of newton-Raphson iterations. This patch
makes this process more streamlined and expose refine_rec has a
way for user to control the applciation of more Newton-Raphson
steps if they feel the need for it.

As an addition, refine_rec allow us to implement fast_rec in scalar
mode for double whenever a SIMD extension allow it. Precision is
around 16ULP for a speed-upof 30-40%