adding some possibilities to exhaustive tests
Note that the old exhaustive will still work. The possibility added is to use two numbers to the command line
to fix the interval in which the exhaustive test will be taken.
Also the outputs have been modifyied: some precision about the test and the ranges in which each ulp tested
are obtained. The only exhaustive tests updated are log exp for scalar and simd